Added ListenerResponse
together with ByteResponse, FileResponse and StreamResponse, they are used to return data to the client, now because of this you can mix different return types, unlike before when you could return either ONLY a byte array or ONLY a stream.

It is also possible to create your own Response types by extending ListenerResponse and overriding one of 2 GetResponse methods, this could be useful for eg. just passing in a class and sending it as Json without writing a lot of Json serializing code.